refactor(interpreter): split enum kind and state-machine statements
